refactor: Split Stages file into MenuStages and GameStages
